Create a new module
Sec. 6, video 37 of Collaboration and Crawling W/ Go

Ref: https://blog.golang.org/using-go-modules

First, we create a repo on GitHub, which is golang-crawling.

Go modules gives us our name spacing. It tells our project where our
code exists. One module to one repo is a good practice.
We copy the code from the reference above.

Run the test with following command:
$ go test

It will tell us what it thinks this particular package is called.

Now we go to the root of our git repo, in this case is golang-crawling,
and run the following command:
$ go mod init <the-namespace-of-this-particular-package>

In this case, it would be github.com/chhsu0222/golang-crawling. Note
that do not include https and be careful not in some subfolder.
The rule of thumb is there should be some way to link the namespace to
where we can find the code on the internet. So which is why the URL is a
good idea.

Note that we had GOPATH and the src folder. In the src folder we had
github.com, the git account and the repo name. That was the namespace
for all the code that was in there.
Now anywhere on my computer I can create code in the folder. Right in
the root of that folder we just do 'go mod init <namespace>'. Now it
knows that everything we are doing here is connected to that repo, that
namespace.
Now if we go down to the folder and run the test again, we can see the
path to the package is now a more general, follows the namespace and
then the subfolders within there.