Add a dependency
Sec. 6, video 38 of Collaboration and Crawling W/ Go

The primary motivation for Go modules was to improve the experience of
using (that is, adding a dependency on) code written by other developers
.

Now we update our hello.go to import rsc.io/quote and use it to
implement Hello function.

Then we run the test again. The first we add a dependency, our local
machine needs to go out and get that code and bring it back. By default
it will try to find the latest stable released.
With Go modules, the major version is a part of the import path. Because
we don't have the version number in our import, it goes to version 1.

To see all direct and indirect dependencies, we can use the following
command:
$ go list -m all

Note that if we import some package that's not go modules yet, our
module will be responsible for getting its indirect dependencies too.