Upgrade a minor dependency
Sec. 6, video 41 of Collaboration and Crawling W/ Go

Before Go modules, "go get" was to download the code into our GOPATH.
If we are in a Go module, "go get" is how we adjust our versions in our
go.mod as well as downloading and pre-downloading the file. Of course we
can manually go into the go.mod file and edit it if we wish. But we
should probably avoid that when "go get" does it better. It will go to
the latest version and following those rules.
$ go get <package-name>

In this case, we upgrade the version of golang.org/x/text
Note that the original version of golang.org/x/text did not require
'tools' package but the new version does require tools.
go.mod would be updated automatically.

Note that the 'text' package is added as an indirect dependency. This
basically states to Go module our code is not directly importing it but
it needs to this particular version to work anyway. So 'rsc.io/quote' or
one of its sub dependencies requires 'text' but we're saying we want
'text' to be this version at minimum.
Make sure that after we upgrade the version of the package, we still
always pass the test.