Upgrade a major dependency
Sec. 6, video 44 of Collaboration and Crawling W/ Go

We can look at the documentation of the package using "go doc", which
takes the path to a package specifically:
$ go doc <package-name>
e.g. go doc rsc.io/quote/v3

We can use "go doc" further:
$ go doc rsc.io/quote/v3.HelloV3

Note that the function name "Hello" was replaced with "HelloV3"
(incompatible).
So we upgrade the version of the package.

Run the test to make sure we didn't break the functionality.
Note that the go.mod was updated.