Make a conversation request
Sec. 11, video 67 of Collaboration and Crawling W/ Go

Ref: https://blog.golang.org/go1.13-errors

In the 2nd solution, the response from Twitter has "items_html" and we
use a library called goquery to parse that.
Note that the "site" of the 2nd solution is the API call from Twitter
instead of the URL for the page in the 1st solution.
goquery is basically a search library built on top of Go's net/html
package and the CSS Selector library cascadia.

We organized our code into 3 functions and a main function.
1st function (makeConversionRequest):
    makes a single request to the Twitter API.
2nd function:
    makes all the requests needed to get all of the items using the 1st
    function
3rd function:
    parse the "items_html" into tweets. The 1st function makes a single
    request to the API. The 2nd function makes all the requests needed
    to get all of the items using the 1st function.

In the 1st function, it takes in the what the current max position is.
It will return a conversation response.
Note that the conversationResponse struct has the json tags to match the
name exactly to the json that they are responding with.
We use url.Values to encode the query parameters to URL properly.
Based on Daniel's test, the max position should be an empty string for
the very first time we make the call to this end point.
Note that whenever we get an HTTP response, make sure we close the
response body.