Make all conversation requests
Sec. 11, video 68 of Collaboration and Crawling W/ Go

In the 2nd function (getConversation), we basically call the 1st
function (makeConversionRequest) multiple times until we have all the
information we need.
We create a variable called "continueCode" to manage the "max_postion"
and "min_position". It's a code that Twitter gives us to continue the
request for more conversation parts. Twitter gives us a "min_position"
in response and we use that as "max_postion" in the next request.
Note that it starts off as an empty string.
If it has more tweets, we continue sending the requests.
We return a slice of string full of all the "items_html" we get.

Note that the robots.txt says that wait 1 second between successive
requests.