Parse HTML
Sec. 11, video 69 of Collaboration and Crawling W/ Go

The 3rd function (parseHtml) reads one of those messages ("items_html")
and returns a slice of tweets, all the tweets it found in that chunk of
HTML.
Golang has a package to parse HTML
(https://godoc.org/golang.org/x/net/html), but it doesn't have anything
to search through it.
This function uses the goquery library
(https://github.com/PuerkitoBio/goquery) to actually examine within that
HTML.
Because the goquery needs to get the data to parse the document from a
Reader, we use strings.NewReader function to convert the message into a
string reader.
Document represents an HTML document to be manipulated.
Because the Colly uses goquery, the syntax looks the same as the 1st
solution.
Through the document, we find the Selection with "tweet" class.
Selection represents a collection of nodes matching some criteria. The
initial Selection can be created by using Document.Find, and then
manipulated using the jQuery-like chainable syntax and methods.

Note that the methods of the *Selection in Document is promoted so we
can use doc.Find directly.